requirements.txt
pypixelcolor>=0.4.0
bleak>=0.21.0
pillow>=10.0.0
yfinance>=0.2.0
google-api-python-client>=2.0.0
requests>=2.31.0
numpy>=1.24.0
